Add and display statistics!
