--- KDH0AZ/BlackjackGame/Card.cs
@@ -1,13 +1,0 @@
-public class Card
-{
-    public string Suit { get; }
-    public string Rank { get; }
-    public int Value { get; }
-
-    public Card(string suit, string rank, int value)
-    {
-        Suit = suit;
-        Rank = rank;
-        Value = value;
-    }
-}
