--- KDH0AZ/BlackjackGame/Dealer.cs
@@ -1,17 +1,0 @@
-using System.Collections.Generic;
-using System;
-
-public class Dealer
-{
-    public List<Card> Hand { get; }
-
-    public Dealer()
-    {
-        Hand = new List<Card>();
-    }
-
-    public void ResetHand()
-    {
-        Hand.Clear();
-    }
-}
