+++ KDH0AZ/BlackjackGame/Models/Card.cs
@@ -1,0 +1,16 @@
+namespace Blackjack.Models
+{
+    public class Card
+    {
+        public string Suit { get; }
+        public string Rank { get; }
+        public int Value { get; }
+
+        public Card(string suit, string rank, int value)
+        {
+            Suit = suit;
+            Rank = rank;
+            Value = value;
+        }
+    }
+}
