+++ KDH0AZ/BlackjackGame/Models/Dealer.cs
@@ -1,0 +1,20 @@
+using System.Collections.Generic;
+using System;
+
+namespace Blackjack.Models
+{
+    public class Dealer
+    {
+        public List<Card> Hand { get; }
+
+        public Dealer()
+        {
+            Hand = new List<Card>();
+        }
+
+        public void ResetHand()
+        {
+            Hand.Clear();
+        }
+    }
+}
