+++ KDH0AZ/BlackjackGame/data.json
@@ -1,0 +1,14 @@
+[
+  {
+    "Name": "Vivi",
+    "Money": 110
+  },
+  {
+    "Name": "Zoli",
+    "Money": 50
+  },
+  {
+    "Name": "Zoli",
+    "Money": 150
+  }
+]
